Fixed problem with trascantion in ReviewService. Integration test were added
