+++ src/main/java/by/gexateq/simplerestservice/service/impl/ReviewServiceImpl.java
@@ -7,6 +7,7 @@ import by.gexateq.simplerestservice.repository.ReviewRepository;
 import by.gexateq.simplerestservice.service.ReviewService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,9 +20,8 @@ import java.util.List;
 @AllArgsConstructor
 public class ReviewServiceImpl implements ReviewService {
 
+    private final ReviewServiceHelper helper;
     private final ReviewRepository reviewRepository;
-
-    private final EmployeeRepository employeeRepository;
 
     @Override
     public List<Review> getReviewsByEmployeeId(Long id) {
@@ -35,7 +35,8 @@ public class ReviewServiceImpl implements ReviewService {
 
         // transactional functional.
         // todo find the problem, fix and provide a test.
-        checkAndUpdateReviewsAndEmployees();
+
+        helper.checkAndUpdateReviewsAndEmployees();
 
         // other non-transactional functional
         // ....
@@ -43,33 +44,5 @@ public class ReviewServiceImpl implements ReviewService {
         log.debug("checkAndUpdateReviews finished");
     }
 
-    @Transactional
-    public void checkAndUpdateReviewsAndEmployees() {
-        LocalDateTime currentDate = LocalDateTime.now();
-        LocalDateTime threeDaysAgo = currentDate.minusDays(3);
-        var reviewsToCancel = reviewRepository.
-                findByCreatedAtBeforeAndStatusNot(threeDaysAgo, ReviewStatus.CANCELLED);
-        log.debug("checkAndUpdateReviews: Found {} reviews to cancel", reviewsToCancel.size());
 
-        List<Review> reviewsToUpdate = reviewsToCancel.stream()
-                .peek(review -> review.setStatus(ReviewStatus.CANCELLED))
-                .toList();
-        if (!reviewsToUpdate.isEmpty()) {
-            reviewRepository.saveAll(reviewsToUpdate);
-            log.debug("checkAndUpdateReviews: Updated {} reviews", reviewsToUpdate.size());
-        }
-
-        var employeesWithAllReviewsCancelled = employeeRepository.findEmployeesWithAllReviewsCancelled();
-        log.debug("checkAndUpdateReviews: Found {} employees to be transferred to inactive status",
-                employeesWithAllReviewsCancelled.size());
-
-        var employeesToUpdate = employeesWithAllReviewsCancelled.stream()
-                .peek(employee -> employee.setIsActive(false))
-                .toList();
-        if (!employeesToUpdate.isEmpty()) {
-            employeeRepository.saveAll(employeesToUpdate);
-            log.debug("checkAndUpdateReviews: Updated {} employees", employeesToUpdate.size());
-        }
-
-    }
 }
