+++ src/main/resources/application.properties
@@ -17,4 +17,7 @@ spring.jpa.properties.hibernate.transaction.jta.platform = none
 #logging.level.org.hibernate.validator=DEBUG
 #logging.level.root=DEBUG
 
+#disable migrations
+spring.flyway.enabled=false
+
 cron.job.check-update-reviews=0 0 0 * * *
