+++ src/test/resources/SQL/insertReview.sql
@@ -1,0 +1,1 @@
+INSERT INTO reviews (status, created_at, employee_id) VALUES ('DRAFT', '2023-01-01 00:00:00', 1);
